pyproject.toml
[project]
name = "cogni-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyzmq>=27.1.0",
  "aiohttp>=3.13.3",
  "gunicorn>=25.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/runtime"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
